+++ core/src/streams/Streambuffer.cpp
@@ -11,15 +11,15 @@ class CStreamLinear
 {
 public:
     CStreamLinear(IStream::Ptr aStream);
-    CStreamLinear(const void* aBuff, uint64_t aSize);
+    CStreamLinear(const void* aBuff, size_t aSize);
     ~CStreamLinear();
 
-    virtual uint64_t Size() const override;
+    virtual size_t Size() const override;
     virtual const void* GetBuff() const override;
 
 private:
     void* m_Buff;
-    uint64_t m_Size;
+    size_t m_Size;
 };
 
 CStreamLinear::CStreamLinear(IStream::Ptr aStream)
@@ -29,7 +29,7 @@ CStreamLinear::CStreamLinear(IStream::Ptr aStream)
     aStream->Read(0, (uint8_t*)m_Buff, m_Size);
 }
 
-CStreamLinear::CStreamLinear(const void* aBuff, uint64_t aSize)
+CStreamLinear::CStreamLinear(const void* aBuff, size_t aSize)
     :m_Size(aSize)
 {
     m_Buff = malloc(m_Size);
@@ -41,7 +41,7 @@ CStreamLinear::~CStreamLinear()
     free(m_Buff);
 }
 
-uint64_t CStreamLinear::Size() const
+size_t CStreamLinear::Size() const
 {
     return m_Size;
 }
@@ -51,7 +51,7 @@ const void* CStreamLinear::GetBuff() const
     return m_Buff;
 }
 
-ILinearStream::Ptr CreateLinearBuffer(const void* aBuff, const uint64_t &aSize)
+ILinearStream::Ptr CreateLinearBuffer(const void* aBuff, const size_t &aSize)
 {
     return ILinearStream::Ptr(new CStreamLinear(aBuff, aSize));
 }
@@ -64,19 +64,19 @@ class CStreamLinearWrite
     :public ILinearWriteStream
 {
 public:
-    CStreamLinearWrite(uint64_t aSize);
+    CStreamLinearWrite(size_t aSize);
     ~CStreamLinearWrite();
 
-    virtual uint64_t Size() const override;
+    virtual size_t Size() const override;
     virtual const void* GetBuff() const override;
     virtual void* GetBuff() override;
 
 private:
     void* m_Buff;
-    uint64_t m_Size;
+    size_t m_Size;
 };
 
-CStreamLinearWrite::CStreamLinearWrite(uint64_t aSize)
+CStreamLinearWrite::CStreamLinearWrite(size_t aSize)
     :m_Size(aSize)
 {
     m_Buff = malloc(m_Size);
@@ -88,7 +88,7 @@ CStreamLinearWrite::~CStreamLinearWrite()
     free(m_Buff);
 }
 
-uint64_t CStreamLinearWrite::Size() const
+size_t CStreamLinearWrite::Size() const
 {
     return m_Size;
 }
@@ -103,7 +103,7 @@ void* CStreamLinearWrite::GetBuff()
     return m_Buff;
 }
 
-ILinearWriteStream::Ptr CreateLinearWriteBuffer(const uint64_t &aSize)
+ILinearWriteStream::Ptr CreateLinearWriteBuffer(const size_t &aSize)
 {
     return ILinearWriteStream::Ptr(new CStreamLinearWrite(aSize));
 }
@@ -113,16 +113,16 @@ class CStreamBuffer
 {
 public:
     CStreamBuffer(IStream::Ptr aStream);
-    CStreamBuffer(const void* aBuff, uint64_t aSize);
+    CStreamBuffer(const void* aBuff, size_t aSize);
     ~CStreamBuffer();
 
-    virtual uint64_t Size() const override;
-    virtual size_t Read(uint64_t Offset, uint8_t* buffer, size_t size) const override;
-    virtual IStream::Ptr Read(uint64_t Offset, size_t size) const override;
+    virtual size_t Size() const override;
+    virtual size_t Read(size_t Offset, uint8_t* buffer, size_t size) const override;
+    virtual IStream::Ptr Read(size_t Offset, size_t size) const override;
 
 private:
     void* m_Buff;
-    uint64_t m_Size;
+    size_t m_Size;
 };
 
 CStreamBuffer::CStreamBuffer(IStream::Ptr aStream)
@@ -132,7 +132,7 @@ CStreamBuffer::CStreamBuffer(IStream::Ptr aStream)
     aStream->Read(0, (uint8_t*)m_Buff, m_Size);
 }
 
-CStreamBuffer::CStreamBuffer(const void* aBuff, uint64_t aSize)
+CStreamBuffer::CStreamBuffer(const void* aBuff, size_t aSize)
     :m_Size(aSize)
 {
     m_Buff = malloc(m_Size);
@@ -144,12 +144,12 @@ CStreamBuffer::~CStreamBuffer()
     free(m_Buff);
 }
 
-uint64_t CStreamBuffer::Size() const
+size_t CStreamBuffer::Size() const
 {
     return m_Size;
 }
 
-size_t CStreamBuffer::Read(uint64_t Offset, uint8_t* buffer, size_t size) const
+size_t CStreamBuffer::Read(size_t Offset, uint8_t* buffer, size_t size) const
 {
     size_t lReadSize = 0;
     if (m_Size > Offset)
@@ -162,7 +162,7 @@ size_t CStreamBuffer::Read(uint64_t Offset, uint8_t* buffer, size_t size) const
     return lReadSize;
 }
 
-IStream::Ptr CStreamBuffer::Read(uint64_t Offset, size_t size) const
+IStream::Ptr CStreamBuffer::Read(size_t Offset, size_t size) const
 {
     IStream::Ptr lStreamBuffer;
     if (m_Size > Offset)
@@ -174,7 +174,7 @@ IStream::Ptr CStreamBuffer::Read(uint64_t Offset, size_t size) const
     return lStreamBuffer;
 }
 
-IStream::Ptr CreateStreamBuffer(const void* aBuff, const uint64_t &aSize)
+IStream::Ptr CreateStreamBuffer(const void* aBuff, const size_t &aSize)
 {
     return IStream::Ptr(new CStreamBuffer(aBuff, aSize));
 }
